// src/types/app/email.ts
export type FieldMenuItems = {
    src: string
    name: string
    value: string
}
// ...
export type MailComposeType = {
    mdAbove: boolean
    composeOpen: boolean
    toggleComposeOpen: () => void
    composePopupWidth: number | string
}